+++ emborg/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '1.8.0'
-__released__ = '2019-10-12'
+__version__ = '1.8.1'
+__released__ = '2019-11-06'
 
 from .settings import Settings as Emborg